+++ app/src/main/java/com/icecrown/onyxridgecm/fragments/CreateProjectHoursTotalFragment.java
@@ -1,0 +1,18 @@
+//**************************************************************
+// Project: OnyxRidge Construction Management
+//
+// File: CreateProjectHoursTotalFragment.java
+//
+// Written by: Raymond O'Neill
+//
+// Written on: 11/20/2020
+//
+// Purpose: Used for logic behind creating a total hours report
+//          for any given project.
+//**************************************************************
+package com.icecrown.onyxridgecm.fragments;
+
+import androidx.fragment.app.Fragment;
+
+public class CreateProjectHoursTotalFragment extends Fragment {
+}